+++ server/front/src/domain/comment/CreateCommentResult.ts
